# smd-vital-backend/docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: SMD Vital - Microservices API
  description: |
    API completa para la plataforma de servicios médicos domiciliarios SMD Vital Bogotá.
    
    ## Servicios Disponibles
    
    - **Authentication Service**: Gestión de autenticación y autorización
    - **User Service**: Gestión de usuarios y perfiles
    - **Appointment Service**: Programación y gestión de citas médicas
    - **Notification Service**: Envío de notificaciones multicanal
    - **Medical Records Service**: Historiales médicos y documentos
    - **Payment Service**: Procesamiento de pagos y facturación
    
    ## Autenticación
    
    Todos los endpoints (excepto públicos) requieren autenticación JWT.
    Incluye el token en el header: `Authorization: Bearer <token>`
    
    ## Códigos de Estado
    
    - `200` - Éxito
    - `201` - Creado exitosamente
    - `400` - Error en la solicitud
    - `401` - No autorizado
    - `403` - Prohibido
    - `404` - No encontrado
    - `422` - Error de validación
    - `500` - Error interno del servidor
  version: 1.0.0
  contact:
    name: SMD Vital Backend Team
    email: dev@smdvitalbogota.com
    url: https://smdvitalbogota.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Desarrollo Local (API Gateway)
  - url: http://localhost:8001
    description: Auth Service Directo
  - url: http://localhost:8002
    description: User Service Directo
...
paths:
...
  /payments:
    post:
      tags:
        - Payments
      summary: Procesar pago
      description: Procesar un nuevo pago para un servicio médico
      operationId: createPayment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreate'
      responses:
        '201':
          description: Pago procesado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /payments/patient/{patientId}:
    get:
      tags:
        - Payments
      summary: Obtener pagos del paciente
      description: Obtener historial de pagos de un paciente
      operationId: getPatientPayments
      security:
        - BearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          description: ID del paciente
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filtrar por estado del pago
          schema:
            type: string
            enum: [pending, completed, failed, refunded]
      responses:
        '200':
          description: Lista de pagos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
  /payments/{paymentId}:
    get:
      tags:
        - Payments
      summary: Obtener pago específico
      description: Obtener detalles de un pago específico
      operationId: getPayment
      security:
        - BearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          description: ID del pago
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalles del pago
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /invoices:
    post:
      tags:
        - Payments
      summary: Crear factura
      description: Crear una nueva factura para un paciente
      operationId: createInvoice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreate'
      responses:
        '201':
          description: Factura creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
  /invoices/patient/{patientId}:
    get:
      tags:
        - Payments
      summary: Obtener facturas del paciente
      description: Obtener facturas de un paciente específico
      operationId: getPatientInvoices
      security:
        - BearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          description: ID del paciente
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filtrar por estado de la factura
          schema:
            type: string
            enum: [pending, paid, overdue, cancelled]
      responses:
        '200':
          description: Lista de facturas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
  /payment-methods:
    post:
      tags:
        - Payments
      summary: Registrar método de pago
      description: Registrar un nuevo método de pago para el usuario
      operationId: createPaymentMethod
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMethodCreate'
      responses:
        '201':
          description: Método de pago registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /payment-methods/user/{userId}:
    get:
      tags:
        - Payments
      summary: Obtener métodos de pago del usuario
      description: Obtener métodos de pago registrados por un usuario
      operationId: getUserPaymentMethods
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID del usuario
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de métodos de pago
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentMethod'
  /refunds:
    post:
      tags:
        - Payments
      summary: Solicitar reembolso
      description: Solicitar reembolso de un pago realizado
      operationId: requestRefund
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequestCreate'
      responses:
        '201':
          description: Solicitud de reembolso creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /reports/revenue:
    get:
      tags:
        - Payments
      summary: Reporte de ingresos
      description: Obtener reporte de ingresos por período (solo administradores)
      operationId: getRevenueReport
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          required: true
          description: Fecha de inicio
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          description: Fecha de fin
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Reporte de ingresos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevenueReport'

  # Notifications Service Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenido del endpoint de login
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
          description: Mensaje de éxito
        data:
          type: object
          description: Datos adicionales
...
    PaymentCreate:
      type: object
      required:
        - patient_id
        - amount
        - currency
        - description
      properties:
        patient_id:
          type: string
          format: uuid
          description: ID del paciente
        appointment_id:
          type: string
          format: uuid
          description: ID de la cita (opcional)
        amount:
          type: number
          format: decimal
          minimum: 0.01
          description: Monto del pago
          example: 150000.00
        currency:
          type: string
          enum: [COP, USD]
          default: COP
          description: Moneda del pago
        payment_method_id:
          type: string
          format: uuid
          description: ID del método de pago
        description:
          type: string
          description: Descripción del pago
          example: "Consulta médica domiciliaria"
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID del pago
        patient_id:
          type: string
          format: uuid
          description: ID del paciente
        appointment_id:
          type: string
          format: uuid
          description: ID de la cita
        amount:
          type: number
          format: decimal
          description: Monto pagado
        currency:
          type: string
          description: Moneda
        payment_method_id:
          type: string
          format: uuid
          description: ID del método de pago
        status:
          type: string
          enum: [pending, completed, failed, refunded]
          description: Estado del pago
        transaction_id:
          type: string
          description: ID de la transacción
        payment_date:
          type: string
          format: date-time
          description: Fecha del pago
        description:
          type: string
          description: Descripción del pago
        processed_by:
          type: string
          format: uuid
          description: ID del usuario que procesó
    InvoiceCreate:
      type: object
      required:
        - patient_id
        - subtotal
        - total_amount
        - due_date
        - items
      properties:
        patient_id:
          type: string
          format: uuid
          description: ID del paciente
        appointment_id:
          type: string
          format: uuid
          description: ID de la cita (opcional)
        subtotal:
          type: number
          format: decimal
          description: Subtotal de la factura
        tax_amount:
          type: number
          format: decimal
          description: Monto de impuestos
          default: 0
        total_amount:
          type: number
          format: decimal
          description: Monto total
        currency:
          type: string
          enum: [COP, USD]
          default: COP
          description: Moneda
        due_date:
          type: string
          format: date
          description: Fecha de vencimiento
        description:
          type: string
          description: Descripción de la factura
        items:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
                description: Descripción del item
              quantity:
                type: integer
                minimum: 1
                description: Cantidad
              unit_price:
                type: number
                format: decimal
                description: Precio unitario
              total:
                type: number
                format: decimal
                description: Total del item
    Invoice:
      allOf:
        - $ref: '#/components/schemas/InvoiceCreate'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              description: ID de la factura
            invoice_number:
              type: string
              description: Número de factura
            status:
              type: string
              enum: [pending, paid, overdue, cancelled]
              description: Estado de la factura
            issue_date:
              type: string
              format: date-time
              description: Fecha de emisión
            created_by:
              type: string
              format: uuid
              description: ID del usuario que creó
    PaymentMethodCreate:
      type: object
      required:
        - user_id
        - type
        - provider
        - last_four
      properties:
        user_id:
          type: string
          format: uuid
          description: ID del usuario
        type:
          type: string
          enum: [credit_card, debit_card, bank_transfer, cash]
          description: Tipo de método de pago
        provider:
          type: string
          description: Proveedor (Visa, Mastercard, etc.)
          example: "Visa"
        last_four:
          type: string
          pattern: "^[0-9]{4}$"
          description: Últimos 4 dígitos
          example: "4532"
        expiry_month:
          type: integer
          minimum: 1
          maximum: 12
          description: Mes de expiración
        expiry_year:
          type: integer
          minimum: 2024
          maximum: 2050
          description: Año de expiración
        is_default:
          type: boolean
          default: false
          description: Es el método por defecto
    PaymentMethod:
      allOf:
        - $ref: '#/components/schemas/PaymentMethodCreate'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              description: ID del método de pago
            is_active:
              type: boolean
              description: Está activo
            created_at:
              type: string
              format: date-time
              description: Fecha de creación
    RefundRequestCreate:
      type: object
      required:
        - payment_id
        - amount
        - reason
      properties:
        payment_id:
          type: string
          format: uuid
          description: ID del pago a reembolsar
        amount:
          type: number
          format: decimal
          description: Monto a reembolsar
        reason:
          type: string
          description: Razón del reembolso
          example: "Cancelación del servicio por parte del paciente"
    RevenueReport:
      type: object
      properties:
        period:
          type: object
          properties:
            start_date:
              type: string
              format: date
            end_date:
              type: string
              format: date
        summary:
          type: object
          properties:
            total_revenue:
              type: number
              format: decimal
              description: Ingresos totales
            payment_count:
              type: integer
              description: Número de pagos
            average_payment:
              type: number
              format: decimal
              description: Pago promedio
        by_payment_method:
          type: object
          additionalProperties:
            type: object
            properties:
              count:
                type: integer
              amount:
                type: number
                format: decimal

    # Notification Schemas
...